examples/Python/python-flask-webapp-sample/requirements.txt
Flask==2.3.2
Flask[async]==2.3.2
azure-identity==1.12.0
azure-appconfiguration-provider==1.1.0b2
aiohttp
